Rename API files
