+++ src/shaders.glsl.h
@@ -4,7 +4,7 @@
 
     Generated by sokol-shdc (https://github.com/floooh/sokol-tools)
 
-    Cmdline: sokol-shdc --input assets/shader.glsl --output src/rive/shader.glsl.h --slang glsl330
+    Cmdline: sokol-shdc --input assets/shaders.glsl --output src/shaders.glsl.h --slang glsl330
 
     Overview:
 
@@ -379,7 +379,7 @@ static const char debug_contour_source_glsl330[215] = {
     0x29,0x3b,0x0a,0x7d,0x0a,0x0a,0x00,
 };
 #if !defined(SOKOL_GFX_INCLUDED)
-  #error "Please include sokol_gfx.h before shader.glsl.h"
+  #error "Please include sokol_gfx.h before shaders.glsl.h"
 #endif
 static inline const sg_shader_desc* rive_debug_contour_shader_desc(sg_backend backend) {
   if (backend == SG_BACKEND_GLCORE33) {
